scripts/inference.py
"""
Inference script for Vietnamese voice conversion using FreeVC
"""

import os
import yaml
import torch
import argparse
from pathlib import Path
import soundfile as sf

import sys
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from models.freevc import FreeVC
from utils.audio import AudioProcessor


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file"""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def load_model(checkpoint_path: str, config: dict, device: torch.device) -> FreeVC:
    """
    Load trained FreeVC model from checkpoint.
    
    Args:
        checkpoint_path: Path to model checkpoint
        config: Configuration dictionary
        device: Device to load model on
    Returns:
        Loaded FreeVC model
    """
    model = FreeVC(config).to(device)
    
    if checkpoint_path and os.path.exists(checkpoint_path):
        checkpoint = torch.load(checkpoint_path, map_location=device)
        model.load_state_dict(checkpoint['generator'])
        print(f'Loaded checkpoint from {checkpoint_path}')
    else:
        print('No checkpoint provided or file not found. Using randomly initialized model.')
    
    model.eval()
    return model


def convert_voice(model: FreeVC,
                 source_audio_path: str,
                 target_speaker_audio_path: str,
                 output_path: str,
                 config: dict,
                 device: torch.device):
    """
    Convert voice from source to target speaker.
    
    Args:
        model: FreeVC model
        source_audio_path: Path to source audio
        target_speaker_audio_path: Path to target speaker reference audio
        output_path: Path to save converted audio
        config: Configuration dictionary
        device: Device
    """
    # Initialize audio processor
    audio_processor = AudioProcessor(config)
    
    print(f'Processing source audio: {source_audio_path}')
    source_data = audio_processor.load_and_process(source_audio_path)
    
    print(f'Processing target speaker audio: {target_speaker_audio_path}')
    target_data = audio_processor.load_and_process(target_speaker_audio_path)
    
    # In a real implementation, you would extract WavLM features here
    # For now, we'll create a dummy tensor to demonstrate the structure
    print('\nNote: In a real implementation, WavLM features would be extracted here.')
    print('This requires loading a pre-trained WavLM model.')
    
    # Dummy WavLM features (in real implementation, extract from source audio)
    # Shape: [1, time_steps, 768]
    dummy_wavlm_features = torch.randn(1, 100, 768).to(device)
    
    # Get target speaker mel-spectrogram
    target_mel = target_data['mel'].to(device)
    
    # Perform voice conversion
    print('\nPerforming voice conversion...')
    with torch.no_grad():
        converted_audio = model(dummy_wavlm_features, target_mel)
    
    # Convert to numpy and save
    converted_audio = converted_audio.squeeze().cpu().numpy()
    
    # Save converted audio
    sample_rate = config.get('audio', {}).get('sampling_rate', 22050)
    sf.write(output_path, converted_audio, sample_rate)
    print(f'Converted audio saved to: {output_path}')
    
    # Print Vietnamese-specific information
    vietnamese_config = config.get('vietnamese', {})
    if vietnamese_config.get('preserve_tones', False):
        print('\nVietnamese tone preservation: ENABLED')
        print(f'Number of Vietnamese tones handled: {vietnamese_config.get("num_tones", 6)}')


def main():
    parser = argparse.ArgumentParser(
        description='Vietnamese voice conversion inference using FreeVC'
    )
    parser.add_argument('--config', type=str, default='config.yaml',
                       help='Path to configuration file')
    parser.add_argument('--checkpoint', type=str, required=False,
                       help='Path to model checkpoint')
    parser.add_argument('--source', type=str, required=True,
                       help='Path to source audio file')
    parser.add_argument('--target', type=str, required=True,
                       help='Path to target speaker reference audio')
    parser.add_argument('--output', type=str, default='converted.wav',
                       help='Path to save converted audio')
    args = parser.parse_args()
    
    # Load config
    config = load_config(args.config)
    
    # Setup device
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    print(f'Using device: {device}')
    
    # Load model
    print('\nLoading FreeVC model with Vietnamese-optimized loss...')
    model = load_model(args.checkpoint, config, device)
    
    # Print loss configuration
    loss_config = config.get('loss', {})
    print('\nOptimized loss configuration for Vietnamese:')
    print(f'  - Reconstruction weight: {loss_config.get("reconstruction_weight", 45.0)}')
    print(f'  - Mel-spectrogram weight: {loss_config.get("mel_loss_weight", 50.0)}')
    print(f'  - Feature matching weight: {loss_config.get("feature_matching_weight", 3.0)}')
    print(f'  - F0 weight (tone preservation): {loss_config.get("f0_loss_weight", 15.0)}')
    print(f'  - Prosody weight: {loss_config.get("prosody_loss_weight", 5.0)}')
    
    # Perform voice conversion
    convert_voice(
        model,
        args.source,
        args.target,
        args.output,
        config,
        device
    )
    
    print('\nVoice conversion completed successfully!')
    print('\nKey features for Vietnamese voice conversion:')
    print('  ✓ Optimized loss weights for Vietnamese tonal preservation')
    print('  ✓ F0 loss for accurate tone reproduction')
    print('  ✓ Prosody loss for natural Vietnamese speech patterns')
    print('  ✓ Enhanced mel-spectrogram loss for better acoustic quality')


if __name__ == '__main__':
    main()
